Move around the top level files and run prettier
