+++ frontend/src/generate-new-doc.js
@@ -1,12 +1,21 @@
 import React from "react";
-import { Button, Container, Row, Col } from 'react-bootstrap';
+import { Button, Container, Row, Col } from "react-bootstrap";
 
 export const GenerateNewPage = () => (
-    <Container className="text-center" style={{display: "flex", alignItems: "center", justifyContent:"center", height: "100vh", width: "100vw" }}>
-      <Row>
-        <Col>
-          <Button variant="primary">Create New Document</Button>
-        </Col>
-      </Row>
-    </Container>
-)
+  <Container
+    className="text-center"
+    style={{
+      display: "flex",
+      alignItems: "center",
+      justifyContent: "center",
+      height: "100vh",
+      width: "100vw"
+    }}
+  >
+    <Row>
+      <Col>
+        <Button variant="primary">Create New Document</Button>
+      </Col>
+    </Row>
+  </Container>
+);
